test(shared): modularize assemblyai client suites
