+++ packages/shared/src/services/__tests__/assemblyai-client/shared/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared type helpers for AssemblyAI client tests.
+ */
+
+export type Listener<TArgs extends unknown[] = unknown[]> = (
+  ...args: TArgs
+) => void;
